+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added `AggOutput`, `SepOutput` and `NullOutput` classes.
+
+### Changed
+
+- Changed `Output` to abstract base class.
+
 ### Configuration
 
 - Added Python 3.12 support.